Add files via upload
COntrol is done by sendig the following format to the Arduino:

[Channel_Name]=[State]{;}

Where:
Channel_Name can be abything fron 1 to 20
State is "ON" or "OFF"
After the state there must be an obligatory ";"

Example:

1=ON;   -> Turning Channel 1 On
14=OFF;   -> Turning Channel 14 Off

Take note that the code is not properly developed and awkward. I have provided comments for some stuff so you can understand what's actually going on.